+++ src/Kmd.Logic.CitizenDocuments.Client/Models/CitizenDocumentUploadRequestModel.cs
@@ -24,21 +24,11 @@ namespace Kmd.Logic.CitizenDocuments.Client.Models
         /// Initializes a new instance of the CitizenDocumentUploadRequestModel
         /// class.
         /// </summary>
-        /// <param name="id"></param>
-        /// <param name="subscriptionId"></param>
-        /// <param name="citizenDocumentConfigId"></param>
-        /// <param name="cpr"></param>
         /// <param name="documentType">Possible values include:
         /// 'CitizenDocument', 'DigitalPostCoverLetter',
         /// 'SnailMailCoverLetter'</param>
-        /// <param name="uploadedAt"></param>
-        /// <param name="uploadedBy"></param>
-        /// <param name="documentUrl"></param>
-        /// <param name="retentionPeriodInDays"></param>
         /// <param name="status">Possible values include: 'InProgress',
         /// 'Completed', 'Failed'</param>
-        /// <param name="fileName"></param>
-        /// <param name="documentName"></param>
         public CitizenDocumentUploadRequestModel(System.Guid? id = default(System.Guid?), System.Guid? subscriptionId = default(System.Guid?), System.Guid? citizenDocumentConfigId = default(System.Guid?), string cpr = default(string), string documentType = default(string), System.DateTime? uploadedAt = default(System.DateTime?), string uploadedBy = default(string), string documentUrl = default(string), int? retentionPeriodInDays = default(int?), string status = default(string), string fileName = default(string), string documentName = default(string))
         {
             Id = id;